renamed package k8s to k8smgmt
