+++ cloud-resource-manager/k8smgmt/appinst.go
@@ -1,4 +1,4 @@
-package k8s
+package k8smgmt
 
 import (
 	"fmt"
