+++ cloud-resource-manager/k8smgmt/helm.go
@@ -1,4 +1,4 @@
-package k8s
+package k8smgmt
 
 import (
 	"fmt"
