+++ cloud-resource-manager/platform/dind/dind-appinst.go
@@ -3,14 +3,14 @@ package dind
 import (
 	"fmt"
 
-	"github.com/mobiledgex/edge-cloud/cloud-resource-manager/k8s"
+	"github.com/mobiledgex/edge-cloud/cloud-resource-manager/k8smgmt"
 	"github.com/mobiledgex/edge-cloud/cloud-resource-manager/platform/pc"
 	"github.com/mobiledgex/edge-cloud/cloudcommon"
 	"github.com/mobiledgex/edge-cloud/edgeproto"
 	"github.com/mobiledgex/edge-cloud/log"
 )
 
-func (s *Platform) CreateAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8s.KubeNames) error {
+func (s *Platform) CreateAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8smgmt.KubeNames) error {
 	log.DebugLog(log.DebugLevelMexos, "call runKubectlCreateApp for dind")
 
 	var err error
@@ -18,9 +18,9 @@ func (s *Platform) CreateAppInst(clusterInst *edgeproto.ClusterInst, app *edgepr
 	appDeploymentType := app.Deployment
 
 	if appDeploymentType == cloudcommon.AppDeploymentTypeKubernetes {
-		err = k8s.CreateAppInst(client, names, app, appInst)
+		err = k8smgmt.CreateAppInst(client, names, app, appInst)
 	} else if appDeploymentType == cloudcommon.AppDeploymentTypeHelm {
-		err = k8s.CreateHelmAppInst(client, names, clusterInst, app, appInst)
+		err = k8smgmt.CreateHelmAppInst(client, names, clusterInst, app, appInst)
 	} else {
 		err = fmt.Errorf("invalid deployment type %s for dind", appDeploymentType)
 	}
@@ -31,7 +31,7 @@ func (s *Platform) CreateAppInst(clusterInst *edgeproto.ClusterInst, app *edgepr
 	return nil
 }
 
-func (s *Platform) DeleteAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8s.KubeNames) error {
+func (s *Platform) DeleteAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8smgmt.KubeNames) error {
 	log.DebugLog(log.DebugLevelMexos, "run kubectl delete app for dind")
 
 	var err error
@@ -39,9 +39,9 @@ func (s *Platform) DeleteAppInst(clusterInst *edgeproto.ClusterInst, app *edgepr
 	appDeploymentType := app.Deployment
 
 	if appDeploymentType == cloudcommon.AppDeploymentTypeKubernetes {
-		err = k8s.DeleteAppInst(client, names, app, appInst)
+		err = k8smgmt.DeleteAppInst(client, names, app, appInst)
 	} else if appDeploymentType == cloudcommon.AppDeploymentTypeHelm {
-		err = k8s.DeleteHelmAppInst(client, names, clusterInst)
+		err = k8smgmt.DeleteHelmAppInst(client, names, clusterInst)
 	} else {
 		err = fmt.Errorf("invalid deployment type %s for dind", appDeploymentType)
 	}
