+++ cloud-resource-manager/platform/platform.go
@@ -1,7 +1,7 @@
 package platform
 
 import (
-	"github.com/mobiledgex/edge-cloud/cloud-resource-manager/k8s"
+	"github.com/mobiledgex/edge-cloud/cloud-resource-manager/k8smgmt"
 	"github.com/mobiledgex/edge-cloud/edgeproto"
 )
 
@@ -19,7 +19,7 @@ type Platform interface {
 	// Delete a Kuberentes Cluster on the cloudlet.
 	DeleteCluster(clusterInst *edgeproto.ClusterInst) error
 	// Create an AppInst on a Cluster
-	CreateAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8s.KubeNames) error
+	CreateAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8smgmt.KubeNames) error
 	// Delete an AppInst on a Cluster
-	DeleteAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8s.KubeNames) error
+	DeleteAppInst(clusterInst *edgeproto.ClusterInst, app *edgeproto.App, appInst *edgeproto.AppInst, names *k8smgmt.KubeNames) error
 }
